requirements.txt
# Meta Marketing API
facebook-business==19.0.3

# Airtable
pyairtable==2.3.3

# Google Sheets
gspread==6.1.2
google-auth==2.29.0

# Utilities
python-dotenv==1.0.1
requests==2.32.3
